services/ai-verification/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
httpx==0.25.1
pillow==10.1.0
face-recognition==1.3.0